Login logic and role check added

+++ src/recipe/ShowRecipe.js
@@ -18,17 +18,17 @@ import { UserContext } from '../App';
 import { useContext } from 'react';
 
 const ShowRecipe = ({ recipe, onDelete }) => {
-    //const {user, login, logout} = useContext(UserContext);
+    const { user, login, logout } = useContext(UserContext);
     const navigate = useNavigate();
 
     const deleteRecipe = async () => {
-        // const user = localStorage.getItem("user");
-        
-            //const u = JSON.parse(user);
+        const user = localStorage.getItem("user");
+        if (user) {
+            const u = JSON.parse(user);
             let response = await fetch(`http://localhost:8080/project/recipe/${recipe.id}`, {
                 method: "DELETE",
                 headers: {
-                    //Authorization: u.token,
+                    Authorization: u.token,
                     "Accept": "application/json",
                     "Content-Type": "application/json",
                 }
@@ -41,6 +41,7 @@ const ShowRecipe = ({ recipe, onDelete }) => {
                 console.log("Error while deleting recipe");
             }
         }
+    }
 
     return (
         <Grid item xs={4}>
@@ -70,12 +71,12 @@ const ShowRecipe = ({ recipe, onDelete }) => {
                         padding: "7% 0%",
                         height: '30px'
                     }}
-                    
+
                     title={recipe.name}
                 />
                 <CardMedia
                     sx={{ height: 140 }}
-                    image ="https://cdn.cancer.ca/-/media/images/stock-images/recipes-main-banner-generic.png?rev=4a884cb602214bb3bc1ce51b84ef8a5f&cx=0.5&cy=0.5&cw=1200&ch=630&hash=B3798D04B58C0F94406103D32F09B1C5"
+                    image="https://cdn.cancer.ca/-/media/images/stock-images/recipes-main-banner-generic.png?rev=4a884cb602214bb3bc1ce51b84ef8a5f&cx=0.5&cy=0.5&cw=1200&ch=630&hash=B3798D04B58C0F94406103D32F09B1C5"
                 />
                 <CardContent
                     sx={{
@@ -96,72 +97,49 @@ const ShowRecipe = ({ recipe, onDelete }) => {
                         <Grid item xs={12}>
                             Time: {recipe.time}
                         </Grid>
-                        {/* <Typography sx={{
-                        marginTop: '10px',
-                        fontStyle: 'italic',
-                        fontFamily: 'RobotoL',
-                        fontWeight: 'normal',
-                        fontSize: '0.8em',
-                        textAlign: 'left',
-                        paddingLeft: '4px',
-                        color: '#156631'
-                    }}>
-                        Info: <br/>
-                        Lorem ipsum dolor sit amet, consectetur adi piscing elit, 
-                        sed magna aliqua sequete.
-                    </Typography> */}
                     </Grid>
                 </CardContent>
 
-                {/* <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
-                {user && user.role === "ROLE_REGULAR_USER" || user.role === "ROLE_CHEF" || user.role === "ROLE_ADMIN" ?
-                    <Tooltip title="Info">
-                        <IconButton
-                            sx={{margin:'0px 8px 15px 8px', color:'#6bb187', fontSize: '1em'}}
-                            aria-label="info"
-                            onClick={() => navigate(`recipe_details/${recipe.ID}`)}>
-                            Informacije&nbsp;&nbsp;<InfoIcon />
-                        </IconButton>
-                    </Tooltip> : 
-                    <> */}
-                <Tooltip title="Info">
-                    <IconButton
-                        sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }}
-                        aria-label="info"
-                        onClick={() => navigate(`recipe_details/${recipe.id}`)}>
-                        <InfoIcon />
-                    </IconButton>
-                </Tooltip>
-                {/*
-                    </>}
-                    {user && user.role === "ROLE_CHEF" ?
-                    <>
-                    <Tooltip title="Edit">
-                        <IconButton sx={{margin:'0px 8px 15px 8px', color:'#6bb187'}} aria-label="edit" onClick={() => navigate(`edit_recipe/${recipe.ID}`)}>
-                            <EditIcon />
-                        </IconButton>
-                    </Tooltip>
-                    <Tooltip title="Delete">
-                        <IconButton sx={{margin:'0px 8px 15px 8px', color:'#6bb187'}} aria-label="delete" onClick={deleteRecipe}>
-                            <DeleteIcon />
-                        </IconButton>
-                    </Tooltip> </> :  <></>}
-                </Box>  */}
-                <Tooltip title="Edit">
-                    <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="edit" onClick={() => navigate(`edit_recipe/${recipe.ID}`)}>
-                        <EditIcon />
-                    </IconButton>
-                </Tooltip>
-                <Tooltip title="Delete">
-                    <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="delete" onClick={deleteRecipe}>
-                        <DeleteIcon />
-                    </IconButton>
-                </Tooltip>
-                <Tooltip title="Add to my Cookbook">
-                    <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="Add to my Cookbook">
-                        <FavoriteBorderIcon />
-                    </IconButton>
-                </Tooltip>
+                <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
+                    {user && user.role === "ROLE_REGULAR_USER" || user.role === "ROLE_CHEF" || user.role === "ROLE_ADMIN" ?
+                        <Tooltip title="Info">
+                            <IconButton
+                                sx={{ margin: '0px 8px 15px 8px', color: '#6bb187', fontSize: '1em' }}
+                                aria-label="info"
+                                onClick={() => navigate(`recipe_details/${recipe.ID}`)}>
+                                Informacije&nbsp;&nbsp;<InfoIcon />
+                            </IconButton>
+                        </Tooltip> :
+                        <>
+                            <Tooltip title="Info">
+                                <IconButton
+                                    sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }}
+                                    aria-label="info"
+                                    onClick={() => navigate(`recipe_details/${recipe.id}`)}>
+                                    <InfoIcon />
+                                </IconButton>
+                            </Tooltip>
+                        </>}
+                    {user && user.role === "ROLE_ADMIN" || user.role === "ROLE_CHEF" ?
+                        <>
+                            <Tooltip title="Edit">
+                                <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="edit" onClick={() => navigate(`edit_recipe/${recipe.ID}`)}>
+                                    <EditIcon />
+                                </IconButton>
+                            </Tooltip>
+                            <Tooltip title="Delete">
+                                <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="delete" onClick={deleteRecipe}>
+                                    <DeleteIcon />
+                                </IconButton>
+                            </Tooltip> </> : <></>}
+                    {user && user.role === "ROLE_REGULAR_USER" ?
+                        <>
+                            <Tooltip title="Add to my Cookbook">
+                                <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="Add to my Cookbook">
+                                    <FavoriteBorderIcon />
+                                </IconButton>
+                            </Tooltip> </> : <></>}
+                </ Box>
             </Card>
         </Grid>
     )
